backend/app/requirements.txt
fastapi==0.111.0
pydantic==1.10.12
python-dotenv==1.0.1
SQLAlchemy==2.0.30
pymysql
cryptography